feat: Add ability to get latest motion events from a camera
Why: To be able to create motion sensors by using the cameras for use in Home Assistant automations

How: By using the get_event_list endpoint we can get all the events for a particular device. This allows us to know the last time a motion event was fired for a camera.