Enable login() to be passed a token
This is necessary with 2fa so that an external program can pass back in a token to persist the token across restarts. If a token is passed in bypass the login process (so that it doesn't just end up raising a twofactorauthentication error while in a non-interactive mode). Instead, lets go ahead and just refresh the token if necessary and continue on with life.